ARM: dts: msm: Add absolute bandwidth for MSM8916
In the cases where the cpu does not vote for DDR absolute bandwidth and
the instantaneous bandwidth vote does not fully help at system level use
cases.
Add support to take the qcom,ab-percent value which indicates the
percentage of instantaneous bandwidth to be used for calculating the
average bandwidth for the cpu.

[SYN: Form patch to Samsung dts.]

Change-Id: I016bb980813ca95276c2f5fe0a725d6b40ebda26
Signed-off-by: Taniya Das <tdas@codeaurora.org>