cpufreq: interactive: Clean up redundant max() usage
Avoid having to pass the max() function to see if we use hispeed_freq each
time we're boosted or above hispeed_load.
Instead of having to check the max() between two different values, add a
condition to see if new_freq is below (or equal) hispeed_freq. If the
calculated freq is already above hispeed, there's no need to do anything,
regardless of boosting.